Moved methods where they belong to.
