+++ src/BitfieldTrait.php
@@ -9,113 +9,15 @@ declare(strict_types=1);
 
 namespace Darealfive\Bitfield;
 
-use Darealfive\Bitfield\filter\Filterable;
-use Darealfive\Bitfield\filter\Type;
 use DomainException;
 use BackedEnum;
 use Stringable;
 
 /**
- * Trait BitfieldTrait implements the {@link Flaggable} interface.
- * It manages to store multiple "flags", each represented by a single BIT of an integer value.
- * It offers methods for setting, adding and removing bits via simple bitwise operations.
- * Since every flag is represented by a single BIT, this class only supports n² values, like: 0,1,2,4,8,16,32,64,128...
- *
- * @see https://www.php.net/manual/en/language.operators.bitwise.php
+ * Trait BitfieldTrait provides static functions to handle bit data
  */
 trait BitfieldTrait
 {
-    /**
-     * @var int the whole bitfield, represented by an unsigned <int>. Defaults to 0 meaning no bit is currently set.
-     */
-    private int $_bitfield = 0;
-
-    /**
-     * Gets the current bitfield as <string> representation.
-     *
-     * @return string e.g. 2 => '10'
-     */
-    public function getBinary(): string
-    {
-        return decbin($this->getBitfield());
-    }
-
-    /**
-     * Gets the current bitfield.
-     *
-     * @return int e.g. 2
-     */
-    public function getBitfield(): int
-    {
-        return $this->_bitfield;
-    }
-
-    /**
-     * Sets the bitfield.
-     *
-     * @param int $bitfield
-     *
-     * @return static
-     */
-    public function setBitfield(int $bitfield): static
-    {
-        $this->_bitfield = self::validateBitfield($bitfield, true);
-
-        return $this;
-    }
-
-    /**
-     * Gets the current bits as a list of <int> exponents mapped to either <int> 1 for "high" or <int> 0 for "low".
-     *
-     * @param Filterable|null $filterable optional filter to be applied on the returned bits
-     *
-     * @return iterable<int,int> maps exponent to the value of the bit
-     */
-    public function getBits(Type $type = Type::ARRAY_FILTER_USE_VALUE, ?Filterable $filterable = null): array
-    {
-        $bits = array_map(
-            intval(...),
-            str_split($this->getBinary())
-        );
-
-        return $filterable?->filter($type, ...$bits) ?: $bits;
-    }
-
-    public function setFlag(int|BackedEnum $bit, int|BackedEnum ...$bits): static
-    {
-        return $this->setBitfield(self::sumBits($bit, ...$bits));
-    }
-
-    public function addFlag(int|BackedEnum $bit, int|BackedEnum ...$bits): static
-    {
-        return $this->setFlag($this->getBitfield() | self::sumBits($bit, ...$bits));
-    }
-
-    public function delFlag(int|BackedEnum $bit, int|BackedEnum ...$bits): static
-    {
-        return $this->setFlag($this->getBitfield() & ~self::sumBits($bit, ...$bits));
-    }
-
-    public function hasFlag(int|BackedEnum $bit, int|BackedEnum ...$bits): bool
-    {
-        return ($this->getBitfield() & self::sumBits($bit, ...$bits)) !== 0;
-    }
-
-    /**
-     * Checks whether all particular bits are set.
-     *
-     * @param int|BackedEnum $bit     a bit to be checked
-     * @param int|BackedEnum ...$bits additional bits to be checked
-     *
-     * @return bool <TRUE> if all provided bits are set, <FALSE> otherwise.
-     */
-    public function hasFlags(int|BackedEnum $bit, int|BackedEnum ...$bits): bool
-    {
-        $bits[] = $bit;
-
-        return count($bits) === count(array_filter($bits, $this->hasFlag(...)));
-    }
-
     /**
      * Sum of given bits. Because we treat arguments as real "bits", this method only sums unique values.
      * E.g.:
